Add non-interactive mode to wallet-cli
