+++ wallet/wallet-cli/Cargo.toml
@@ -14,6 +14,7 @@ wallet-controller = { path = "../wallet-controller" }
 wallet-storage = { path = '../storage' }
 
 clap = { version = "4", features = ["derive"] }
+crossterm = "0.26"
 directories = "5.0"
 env_logger = "0.10"
 reedline = { version = "0.19", features = ["external_printer"] }
