+++ wallet/wallet-cli/src/repl/interactive/log.rs
