+++ wallet/wallet-cli/src/repl/interactive/wallet_completions.rs
