+++ wallet/wallet-cli/src/repl/interactive/wallet_prompt.rs
